requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.3
Pillow==10.1.0
python-multipart==0.0.6
scikit-learn==1.3.2
python-dotenv==1.0.0
cors==1.0.1 